requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.0.6
django-cors-headers==3.2.0
django-crispy-forms==1.9.2
django-filter==2.2.0
django-filters==0.2.1
django-formtools==2.2
django-import-export==2.4.0
django-redis==4.11.0
django-reversion==3.0.8
djangorestframework==3.10.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
PyJWT==1.7.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.25.11
xadmin @ https://codeload.github.com/sshwsfc/xadmin/zip/django2
xlrd==1.2.0
xlwt==1.3.0
